Organise into packages (#15)
* Add duke as top level package

* Split classes into relevant packages

* Remove unused variables
