+++ src/main/java/duke/Duke.java
@@ -1,11 +1,13 @@
-import java.util.Scanner;
+package duke;
+
+import duke.command.Command;
+import duke.exception.DukeException;
+import duke.parser.Parser;
+import duke.storage.Storage;
+import duke.task.TaskManager;
+import duke.ui.Ui;
 
 public class Duke {
-    // User-facing error messages
-    private static final String INVALID_COMMAND_MESSAGE = "Invalid command.";
-    private static final String INVALID_TASK_NUMBER_MESSAGE = "Invalid task number.";
-
-    private final Scanner scanner = new Scanner(System.in);
     private final Storage storage;
     private TaskManager taskManager;
     private final Ui ui;
