+++ src/main/java/duke/command/AddToDoCommand.java
@@ -1,3 +1,11 @@
+package duke.command;
+
+import duke.exception.DukeException;
+import duke.storage.Storage;
+import duke.task.TaskManager;
+import duke.task.ToDo;
+import duke.ui.Ui;
+
 public class AddToDoCommand extends Command {
     public static final String COMMAND_WORD = "todo";
     public static final String USAGE_MESSAGE = "To add a new todo, use 'todo <name>'.";
