+++ src/main/java/duke/command/ExitCommand.java
@@ -1,3 +1,9 @@
+package duke.command;
+
+import duke.storage.Storage;
+import duke.task.TaskManager;
+import duke.ui.Ui;
+
 public class ExitCommand extends Command {
     public static final String COMMAND_WORD = "bye";
     public static final String USAGE_MESSAGE = "To close Duke, use 'bye'";
