+++ src/main/java/duke/parser/Parser.java
@@ -1,3 +1,15 @@
+package duke.parser;
+
+import duke.command.AddDeadlineCommand;
+import duke.command.AddEventCommand;
+import duke.command.AddToDoCommand;
+import duke.command.Command;
+import duke.command.CompleteTaskCommand;
+import duke.command.DeleteTaskCommand;
+import duke.command.ExitCommand;
+import duke.command.HelpCommand;
+import duke.command.ListCommand;
+
 public class Parser {
     public static Command parse(String fullCommand) {
         final String trimmedCommand = fullCommand.trim();
