+++ src/main/java/duke/storage/Storage.java
@@ -1,3 +1,8 @@
+package duke.storage;
+
+import duke.exception.DukeException;
+import duke.task.Task;
+
 import java.io.BufferedReader;
 import java.io.File;
 import java.io.FileReader;
