+++ src/main/java/duke/task/Deadline.java
@@ -1,3 +1,8 @@
+package duke.task;
+
+import duke.util.DukeDateTime;
+import duke.exception.DukeException;
+
 public class Deadline extends Task implements Timestampable {
     private final DukeDateTime dueDate;
