+++ src/main/java/duke/task/Timestampable.java
@@ -1,3 +1,7 @@
+package duke.task;
+
+import duke.util.DukeDateTime;
+
 public interface Timestampable {
     public boolean onSameDayAs(DukeDateTime date);
 }
