+++ src/main/java/duke/ui/Ui.java
@@ -1,3 +1,7 @@
+package duke.ui;
+
+import duke.task.TaskManager;
+
 import java.util.Scanner;
 
 public class Ui {
